I hope that this commit has the updates required for the files in dsp_manager (placed directly in src/extensions). The problem is that the code was done before lib_src existed. In the original code, dsp_wrapper includes SSRC.h and then uses what existed in that module. However, there is no such module anymore. In the official lib_src (1.0.0) there is ssrc.h. However, in this commit we are using lib_src from github, https://github.com/xmos/lib_src. Examples for using it, see ../tests/ssrc_test/ssrc_test.xc. Essentially, dsp_wrapper.c includes src.h from ../lib_src/api. Of course, expect to be able to run dsp_manager with the official lib_src.
When doing a build of i0o2xxxxxx there is one error:
"too many arguments to function `buffer'"
This is related to usb_buffer.xc in module_usb_audio, i.e. this is the next step
in the quest for integrating ssrc into usb_audio.

Also observe that I had to trick xscope_int() in src_manager.xc