Add update to usb_buffer.xc for SSRC_DEMO.
Not fully clear how to add SSRC_DEMO code to the SOF notification case. This
since notification code has been changed quite a lot. The old implementation
remains and is guarded with #if 0. However, in order to use the SSRC_DEMO notification
code, the #if 0 has been replaced with SSRC_DEMO, i.e. is activated when code is
compiled with SSRC_DEMO defined.