Add better configuration handling for building ssrc_demo and activation/deactivation of the original SOF feedback in usb_buffer.xc. In practice, have introduced 3 configs in the Makefile that handles SSRC_DEMO and SOF_FEEDBACK_OLD
i0o2xxxxxx
Just a normal build of the usb framework

i0o2xxxxxx_oldsof
Build with activation of the the old SOF feedback. Works without any problems -
important that it works since it's used in the next config with SSRC_DEMO.

i0o2xxxxxx_ssrc_oldsof
Build with old SOF and with SSRC_DEMO code. As noted in previous commits, xscope
investigations show that src_manager_task receives audio from usb host but that
data back from dsp_task are 0. Therefore, we get nothing else than 0 on i2s. This
is also the case when host and i2s are running at the same sample rate. It's
possible to change i2s sample rate (44100 - 192000) with button 1, but this is not
shown on the LED matrix as in the original ssrc demo. More important, SSRC report
failure for all sample rates that are not the same for host and i2s.