Add solution to the i2s problem.
Turned out that sSSRCCtrl structure was not correctly initialized in
dsp_init(), sSSRCCtrl[instance].uiNchannels was never set, i.e. 0.
The result of this was that the for-loop that populated output buffer
piOut was never executed. See for example SSRC_proc_F1_F2().

With the correct init, SSRC_DEMO now runs as intended and can handle
samples rates fromm 44100 to 192000.

Have noticed that xscope and/or printf() can introduce problems when
running samples rates above 96000.